requirements.txt
streamlit
pandas
openai
sentence-transformers
transformers
python-dotenv